feat(UserInformation): renamed password tab on settings to user information for user related content
