+++ apps/Conduit-UI/src/components/navigation/InnerLayouts/settingsLayout.tsx
@@ -3,11 +3,11 @@ import { Settings } from '@mui/icons-material';
 import StyledLayout from './styledLayout';
 
 const SettingsLayout: React.FC = ({ children }) => {
-  const pathNames = ['/settings/settings', '/settings/admins', '/settings/password'];
+  const pathNames = ['/settings/settings', '/settings/admins', '/settings/userSettings'];
   const labels = [
     { name: 'general', id: 'settings' },
     { name: 'admin users', id: 'admins' },
-    { name: 'change password', id: 'password' },
+    { name: 'user settings', id: 'userSettings' },
   ];
 
   return (
